+++ src/modules/main/mnemonic/MnemonicComponent.tsx
@@ -1,6 +1,5 @@
 import React, {FC, useState} from "react";
 import s from "../EngWord/EngWord.module.css";
-import s2 from "./mnemonic.module.css";
 import {IEngWord, IMnemonic} from "../../../shared/models/engWordTypes";
 import HighlightMnemonic from "./HighlightMnemonic";
 import {useAppDispatch} from "../../../store";
@@ -13,6 +12,7 @@ import {
 } from "../../../store/engWordSlice";
 import {ReactComponent as HeartNo} from "../../../import/icons/heart.svg"
 import {ReactComponent as HeartYes} from "../../../import/icons/heart-clicked.svg"
+import {ReactComponent as GreyHeart} from "../../../import/icons/heart-grey.svg"
 import {ReactComponent as Pencil} from "../../../import/icons/pencil.svg"
 import {ReactComponent as Trash} from "../../../import/icons/trash-bin.svg"
 import {ReactComponent as Add} from "../../../import/icons/add.svg"
@@ -23,12 +23,12 @@ import UpdateMnemonic from "./UpdateMnemonic";
 import {DateAgo} from "../DateAgo";
 import {useNavigate} from "react-router";
 import {DeleteModal} from "../Modal/DeleteModal";
+import moment from "moment";
 
 type PropsType = {
     mnemonic: IMnemonic
     auth: boolean
     engWord: IEngWord
-
 }
 const MnemonicComponent: FC<PropsType> = (props) => {
     const [bold, setBold] = useState<boolean>(false);
@@ -41,18 +41,19 @@ const MnemonicComponent: FC<PropsType> = (props) => {
         setShowModal(false)
     }
     const dispatch = useAppDispatch();
-   const navigate = useNavigate()
+    const navigate = useNavigate()
     let addLikeMnemonic = () => {
         if (props.auth) {
-            if (!props.mnemonic.isLiked) {
-                dispatch(addMnemonicLikeAsync(props.mnemonic.mnemonicId))
-            } else {
-                dispatch(deleteMnemonicLikeAsync(props.mnemonic.mnemonicId))
+            if (!props.mnemonic.isCreator) {
+                if (!props.mnemonic.isLiked) {
+                    dispatch(addMnemonicLikeAsync(props.mnemonic.mnemonicId))
+                } else {
+                    dispatch(deleteMnemonicLikeAsync(props.mnemonic.mnemonicId))
+                }
             }
         } else {
             navigate(`/login`)
         }
-
     };
     const showDeleteModal = () => {
         setShowModal(true)
@@ -65,7 +66,7 @@ const MnemonicComponent: FC<PropsType> = (props) => {
     let addMeMnemonic = () => {
         if (!props.mnemonic.added) {
             dispatch(addMeMnemonicAsync(props.mnemonic.mnemonicId))
-        }else {
+        } else {
             dispatch(deleteMeMnemonicAsync(props.mnemonic.mnemonicId))
         }
     };
@@ -83,16 +84,20 @@ const MnemonicComponent: FC<PropsType> = (props) => {
         addDelete = <Minus/>;
     }
     let boldIcon = <NotBold/>;
-        if(bold){
-            boldIcon = <Bold/>
-        }
-        let date = (date: string) => {
-            return (
-                <div>
-                    <DateAgo date={date}/>
-                </div>
-            )
-        }
+    if (bold) {
+        boldIcon = <Bold/>
+    }
+    let date = (date: string) => {
+        return (
+            <div>
+                <DateAgo date={date}/>
+            </div>
+        )
+    }
+    let isExpired = (created: string) => {
+        let createdDate = moment(created, "DD-MM-YYYY HH:mm:ss")
+        return  createdDate.add(24, 'hours').isBefore(moment())
+    }
 
     return (
 
@@ -107,55 +112,64 @@ const MnemonicComponent: FC<PropsType> = (props) => {
                         <div className={s.accuracy}>
                             {Math.round(props.mnemonic.accuracy)}%
                         </div>
-                        <div className={s.dateFormat}>
-                            {date(props.mnemonic.created)}
+
+                        <div className={s.likeIcon}>
+                            <div className={s.likeCountContainer}>
+                                <span className={s.likeCount}>{props.mnemonic.likes}</span>
+                            </div>
+                            {props.mnemonic.isCreator ?
+                                <div>
+                                    <GreyHeart/>
+                                </div>
+                                :
+                                <div onClick={addLikeMnemonic} className={s.myRedHeart}>
+                                    {Like}
+                                </div> }
                         </div>
+
                     </div>
 
-
-                <div className={s.mnemo}>
-                    {bold ? <HighlightMnemonic
-                        highlight={props.mnemonic.highlight}
-                        mnemonic={props.mnemonic.phrase}
-                    /> : props.mnemonic.phrase}
-                </div>
-
-                <div className={s.mnemonicBottom}>
-                    <span className={s.author}>автор: <> </>
-                        {props.mnemonic.creator.nickname}
-                    </span>
-
-
-                    <div className={s.buttonIcons}>
-
-                        <div className={s.boldIconBox}>
-                            <div className={s.prompt}> мнемо-часть </div>
-                            <div onClick={pushBold}>{boldIcon}</div>
+                    <div className={s.mnemo}>
+                        {bold ? <HighlightMnemonic
+                            highlight={props.mnemonic.highlight}
+                            mnemonic={props.mnemonic.phrase}
+                        /> : props.mnemonic.phrase}
+                    </div>
+                    <div className={s.mnemonicBottom}>
+                        <div className={s.creatorInfo}>
+                            <div className={s.author}>
+                                {props.mnemonic.creator.nickname}
+                            </div>
+                            <div className={s.dateFormat}>
+                                {date(props.mnemonic.created)}
+                            </div>
                         </div>
-
-                        {!props.mnemonic.isCreator && props.auth &&
-                        <div onClick={addMeMnemonic}> {addDelete} </div>
-                        }
-                        {props.mnemonic.isCreator && !props.mnemonic.exampleExists && props.auth &&
-                            <div className={s.boldIconBox}>
-                                <div className={s.prompt}> удалить </div>
-                                <div onClick={showDeleteModal} className={s2.deleteIcon}><Trash/></div>
+                        <div className={s.buttonIcons}>
+                            <div className={s.boldIconBox} onClick={pushBold}>
+                                <div  className={s.iconContainer}>{boldIcon}</div>
+                                <div> Выделить </div>
                             </div>
-
-                        }
-                        {props.mnemonic.isCreator && !props.mnemonic.exampleExists && props.auth &&
-                            <div className={s.boldIconBox}>
-                                <div className={s.prompt}> редактировать </div>
-                                <div onClick={editMnemo}> <Pencil/> </div>
-                            </div>
-
-                        }
-                        <div onClick={addLikeMnemonic}>{Like}</div>
-                        <div className={s.likeCountContainer}>
-                            <span className={s.likeCount}>{props.mnemonic.likes}</span>
+                            {!props.mnemonic.isCreator && props.auth && !props.mnemonic.myExampleExists && // поменять флаг на !props.mnemonic.isCreator
+                                <div className={s.boldIconBox} onClick={addMeMnemonic}>
+                                    <div  className={s.iconContainer}> {addDelete} </div>
+                                    <div> Избранное </div>
+                                </div>
+                            }
+                            {props.mnemonic.isCreator && !props.mnemonic.exampleExists && props.auth
+                                && !isExpired(props.mnemonic.created) &&
+                                <div className={s.boldIconBox} onClick={showDeleteModal}>
+                                    <div  className={s.iconContainer}><Trash/></div>
+                                    <div> Удалить </div>
+                                </div>
+                            }
+                            {props.mnemonic.isCreator && !props.mnemonic.exampleExists && props.auth &&
+                                <div className={s.boldIconBox} onClick={editMnemo}>
+                                    <div  className={s.iconContainer}><Pencil/></div>
+                                    <div> Редактировать </div>
+                                </div>
+                            }
                         </div>
                     </div>
-                </div>
                 </>
                 :
                 <>
@@ -163,15 +177,14 @@ const MnemonicComponent: FC<PropsType> = (props) => {
                         mnemonicId={props.mnemonic.mnemonicId}
                         mnemonicPhrase={props.mnemonic.phrase}
                         setEdit={setEdit}
-                        engWord ={props.engWord}
+                        engWord={props.engWord}
                     />
                 </>
             }
+
             <DeleteModal
-                classElement = {"мнемонику"}
-                elementMnemonic={props.mnemonic}
-                elementExample={null}
-                elementMyExample ={null}
+                classElement={"мнемонику"}
+                elementToDelete={props.mnemonic.phrase}
                 show={showModal}
                 close={closeDeleteModal}
                 delete={deleteMnemonic}
